requirements.txt
PyQt6==6.7.0
numpy==2.1.1
sounddevice==0.4.7
soundfile==0.12.1
mido==1.3.2
python-rtmidi==1.5.8
pyttsx3==2.90
PyYAML==6.0.2
rich==13.9.2
